# feat: optimize string processing with memory-safe implementation - 64% faster